chore(refactor): migrate reqwest_error handler
Migrate it to it's own for better readability

This can not be unit testable.
